[*] Corrige o tipo da propriedade cUFOrig e o case na seção origComb na rotina de leitura do ini. Por: WesleySantos https://www.projetoacbr.com.br/forum/topic/74923-tag-origcomb-no-acbrlib-obrigat%C3%B3rio-o-preenchimento-do-grupo-de-uf-de-origem-de-combustivel/?do=findComment&comment=484446&_rid=124795
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@31146 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Projetos/ACBrLib/Demos/C#/NFe/Imports/Dinamico/Shared/NotaFiscal/origComb.cs
@@ -7,7 +7,7 @@
 
         public IndImport indImport { get; set; }
 
-        public decimal cUFOrig { get; set; }
+        public int cUFOrig { get; set; }
 
         public decimal pOrig { get; set; }
